+++ example/lib/main.dart
@@ -41,8 +41,9 @@ class _ZonkaFeedBackSurveyState extends State<ZonkaFeedBackSurvey>
         context: context,
         displayType: 'slide-up',
         minimumHeight: 410,
-        closeIconPosition: "right",
+        closeIconPosition: "left",
         autoClose: true,
+        closeIconType: 'icon2',
         expandedHeight: 580);
 
     WidgetsBinding.instance.addObserver(this);